Custom hook for useNavigate related functions created
