+++ src/pages/genre/book-list/mobile-view.tsx
@@ -1,5 +1,5 @@
-import { useEffect } from "react";
-import { useNavigate, useParams } from "react-router-dom";
+import { useEffect, useMemo } from "react";
+import { useParams } from "react-router-dom";
 import { useAppDispatch, useAppSelector } from "../../../redux/hooks";
 import genresActions from "../../../redux/actions/genres";
 import type { RootState } from "../../../redux/store";
@@ -8,15 +8,16 @@ import { MoreGenresSelect } from "../../../components/genres/more-genres-select"
 import type { GenreBookListRouteParams } from ".";
 import { GenreBookListItemsMobile } from "../../../components/genres/list-items-mobile";
 import { urlSlugToCapitalizedText } from "../../../helpers/utils";
+import { useNavigateTo } from "../../../hooks/navigateTo";
 
 
 export const GenreBookListMobile = () => {
-    const navigate = useNavigate();
     const dispatch = useAppDispatch();
     const params = useParams<GenreBookListRouteParams>();
 
     const { genresList, status: genresStatus } = useAppSelector((state: RootState) => state.genres);
     const { latestReleases, mostRatedBooks, bestRatedBooks, status: editionsStatus } = useAppSelector((state: RootState) => state.editions);
+    const { handleClickOnGenresPage, handleClickOnSelectedGenre } = useNavigateTo(params);
 
     const bookListsMap = {
         "latest-releases": latestReleases,
@@ -24,16 +25,10 @@ export const GenreBookListMobile = () => {
         "best-rated": bestRatedBooks.list,
     };
     const selectedList = (params.list && bookListsMap[params.list]) || []
-    const genreName = urlSlugToCapitalizedText(params.genre || "")
-    const listName = urlSlugToCapitalizedText(params.list || "")
+    const genreName = useMemo(() => urlSlugToCapitalizedText(params.genre || ""), [params.genre]);
+    const listName = useMemo(() => urlSlugToCapitalizedText(params.list || ""), [params.list]);
 
-    const handleClickOnGenresPage = () => {
-        navigate("/genres");
-    }
 
-    const handleClickOnSelectedGenre = () => {
-        navigate(`/genres/${params.genre}`);
-    }
 
     useEffect(() => {
         dispatch(genresActions.getAll({ limit: 20, sortBy: "occurrence" }));
